% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/border_data.r
\docType{data}
\name{border}
\alias{border}
\title{Majete study area borders}
\format{Simple feature polygon}
\usage{
data("border")
}
\description{
This data-set contains the borders for Majete \bold{\emph{focal area A}}, relating to the study of the prevalence of malaria in Chikhwawa district, southern Malawi.
The data-set contains Geometry set for 1 feature.

\itemize{
  \item \code{Geometry type:} Polygon.
  \item \code{dimension:} XY.
  \item \code{bbox:} xmin: 654.6224 ymin: 8243.117 xmax: 664.3984 ymax: 8253.008
  \item \code{epsg (SRID):} 32736
  \item \code{proj4string:} +proj=utm +zone=36 +south +datum=WGS84 +units=m +no_defs
}
}
\keyword{datasets}
