% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majete_data.r
\docType{data}
\name{majete}
\alias{majete}
\title{Majete malaria prevalence data}
\format{A data frame with 747 features and 7 variables}
\usage{
data("majete")
}
\description{
This data-set relates to malaria prevalence study conduted in Majete (Chikwawa), southern Malawi.
The variables are as follows:

\itemize{
  \item \code{rdt:} Rapid diagnostic test result;  0 = negative, 1 = positive.
  \item \code{age:} Age of the individual in months.
  \item \code{quintile:} Wealth quintile; ranging from 1 = poor to 5 = well to do.
  \item \code{itn:} Insecticide treated bed-net usage; 0 = no, 1 = yes.
  \item \code{elev:} Elevation; height above sea level in meters.
  \item \code{ndvi:} Normalised difference vegetation index (greenness).
  \item \code{agecat:} Age category; 1 = child, 2 = adult.
  \item \code{geometry:} Point or household locations (UTM).
}
}
\references{
Kabaghe A N, Chipeta M G, McCann R S, Phiri K S, Van Vugt M, Takken W, Diggle P J, and Terlouw D J. (2017). Adaptive geostatistical sampling enables efficient identification of malaria hotspots in repeated cross-sectional surveys in rural Malawi, \emph{PLoS One} \bold{12}(2) pp. e0172266
}
\keyword{datasets}
